MessageWindowLogger: make thread safe
We dont want to have to start making everything thread safe (core, GameControl, DisplayMessage, etc) so I'm opting
to handle MessageWindow logging as a special case on the main thread
